Feature/LRA-213 build tools sfcc ci

* LRA-213 #comment first commit for refactor of build_tools upload

    Removes a lot of unused upload code
    Removes the upload and uploadCartridge commands
    Moves tasks into a /tasks folder
    Moves shared code into a /lib folder

* LRA-213 authentication and code activation via SFCC-CI

* LRA-213 code deployment via SFCC-CI

* LRA-213 data deployment via SFCC-CI

* LRA-213 #comment renames npm script 'deploy' to 'deployData'; adjusts readme; adjusts dw.json.example for clientId and secret

* LRA-213 jsdoc comments; improves error handling; improves logging; refactoring to code and data deployment scriptss

* LRA-213 create version properties file

* LRA-213 logs last data deployment timestamp

* LRA-213 commits package-lock.json

* LRA-213 removes the dependency for a root dw.json

* LRA-213 updates readmes; updates dw.json.example; updates contributors

* LRA-213 adds documentation on 'npm run clean' command

* LRA-213 #comment changes naming for 'On-Demand Sandboxes' to 'Unified Authentication'

* LRA-213 #comment adjusts package.json to reference LYONSCG internal repo

* LRA-213 resolves package file correctly when calling from builders.xml

* LRA-213 resolves issues with requiring package.json with builders.xml as callee

* Revert "Merge branch 'bugfix/LRA-213-resolve-package-file' into feature/LRA-213-build-tools-sfcc-ci"

    This reverts commit b882d10fbecc0f63471967968808c11b6da937c0, reversing
    changes made to de106eea25f210e6b7463197069cedc0c1929143.

* LRA-213 #comment re-throws errors after proper logging so that jenkins builds can fail

* LRA-213 #comment process.exit with code 1 after proper logging so that jenkins builds can fail

* LRA-213 #comment commits package-lock changes; fixes faulty repo url in package.json; fixes quotes in builders xml; use BUILD_TAG or 'build' as default codeVersion names; fix relative file path issue in upload-utils;

* LRA-213 #comment adjusts readme examples; adds information about continuous integration and account manager; removes duplicate testing info

* LRA-213 #comment adds screenshot of account manager to readme; adds blurb about data deployment logs; adds blurb about sfcc-ci and ssh key

* LRA-213 Fix typos in README.md and new package-lock.json

* LRA-213 #comment adds defaultOptions object that will be used if a parameter is not set in either dw.json or cli args; moves code-version to a defaultOption

* LRA-213 Updated utilities used by postProjectToSARMS to better handle
    the sfcc-ci entry in package.json which has a urn instead of a semver
    n.n.n version. postProjectToSARMS will use "not specified" as the
    version for the sfcc-ci component and will reformat the package.json
    version string as a url which is stored as the repository attribute for
    the module in the SARMS system

Approved-by: Matt Rose <matt.j.rose@capgemini.com>
Approved-by: Gabriel Freiberg <gabefreiberg@gmail.com>
Approved-by: Ryan Walter <rwalter@lyonscg.com>

+++ build_tools/util/sarms/lib/sarms.js
@@ -93,44 +93,56 @@ function createComponentVulnerabilityDescriptions(bomDependencyComponents, packa
  */
 function buildListOfComponentsFromPackage(dependencies, dependenciesKeys, platformName, categoryName, tempNpmReportPath) {
   var re = /(\d+\.)(\d+)(\..*)?/g,
+      versionString = '',
       bomDependencyComponents = {},
       moduleInfo = {};
 
   for (let dependencyKey of dependenciesKeys) {
+    versionString = dependencies[dependencyKey].match(re);
+    versionString == null ? versionString = '' : versionString = versionString[0] + ''; 
     bomDependencyComponents[dependencyKey] = {
-        "name" : dependencyKey,
-        "version" : ((dependencies[dependencyKey]).match(re)).join(""),
-        "platform" : 'NodeJS',
-        "category" : categoryName,
+      "name" : dependencyKey,
+      "version" : versionString,
+      "platform" : 'NodeJS',
+      "category" : categoryName,
     }
-    if (shell.exec('npm view ' 
-                    + bomDependencyComponents[dependencyKey].name 
-                    + '@' + bomDependencyComponents[dependencyKey].version
-                    + ' name description homepage repository bugs license gitHead _id --json > '
-                    + tempNpmReportPath).code == 0) {
-      try {
-        moduleInfo = fileUtils.readJSONFromFile(tempNpmReportPath);
-      } catch (e) {
-        eh.errorHandler('error', e);
+    // If the version in the package.json is not in the semver format (e.g. bitbucket:lyonsconsultinggroup/sfcc-ci.git#master)
+    // but appears to look like a url (contains :// or bitbucket:), treat the string like a repo url and set a default version string
+    if (versionString == '') {
+      bomDependencyComponents[dependencyKey].version = 'not specified';
+      if ((dependencies[dependencyKey].search(/^.+:\/\//) >= 0) || (dependencies[dependencyKey].search(/^bitbucket:/) >= 0)) {
+        bomDependencyComponents[dependencyKey].repository = fixRepoUrl.fixRepoUrl(dependencies[dependencyKey]);
       }
-      
-      fs.unlinkSync(tempNpmReportPath);
-      bomDependencyComponents[dependencyKey].description = moduleInfo.description
-      if (typeof moduleInfo.bugs != 'undefined') {
-        bomDependencyComponents[dependencyKey].description += '\n\nBugs:' + moduleInfo.bugs.url;
+    } else
+      // Otherwise, try to get the details using npm view
+      if (shell.exec('npm view ' 
+                      + bomDependencyComponents[dependencyKey].name 
+                      + '@' + bomDependencyComponents[dependencyKey].version
+                      + ' name description homepage repository bugs license gitHead _id --json > '
+                      + tempNpmReportPath).code == 0) {
+        try {
+          moduleInfo = fileUtils.readJSONFromFile(tempNpmReportPath);
+        } catch (e) {
+          eh.errorHandler('error', e);
+        }
+        
+        fs.unlinkSync(tempNpmReportPath);
+        bomDependencyComponents[dependencyKey].description = moduleInfo.description
+        if (typeof moduleInfo.bugs != 'undefined') {
+          bomDependencyComponents[dependencyKey].description += '\n\nBugs:' + moduleInfo.bugs.url;
+        }
+        if (typeof moduleInfo.gitHead != 'undefined') {
+          bomDependencyComponents[dependencyKey].description += '\n\ngitHead:' + moduleInfo.gitHead;
+        }
+        if (typeof moduleInfo.repository != 'undefined') {
+          bomDependencyComponents[dependencyKey].repository = fixRepoUrl.fixRepoUrl(moduleInfo.repository.url);
+        }
+        if (typeof moduleInfo.homepage != 'undefined') {
+          bomDependencyComponents[dependencyKey].news = moduleInfo.homepage;
+        }
+      } else {
+          eh.errorHandler('warn', 'Unable to execute the ‘npm view’ command for the \'' + dependencyKey + '\' module.');
       }
-      if (typeof moduleInfo.gitHead != 'undefined') {
-        bomDependencyComponents[dependencyKey].description += '\n\ngitHead:' + moduleInfo.gitHead;
-      }
-      if (typeof moduleInfo.repository != 'undefined') {
-        bomDependencyComponents[dependencyKey].repository = fixRepoUrl.fixRepoUrl(moduleInfo.repository.url);
-      }
-      if (typeof moduleInfo.homepage != 'undefined') {
-        bomDependencyComponents[dependencyKey].news = moduleInfo.homepage;
-      }
-    } else {
-        eh.errorHandler('warn', 'Unable to execute the ‘npm view’ command for the \'' + dependencyKey + '\' module.');
-    }
   }
 
   return bomDependencyComponents;
